add default constructor to Pixel, adjust file type to uint8 instead of int for pixel colors, implement read image method (still need debugging)

+++ pixel.cpp
@@ -5,7 +5,7 @@ int* Pixel::getPixel (){
     return wholePixel;
 }
 
-void Pixel::setPixel( int r, int g, int b) {
+void Pixel::setPixel(uint8_t r, uint8_t g, uint8_t b) {
     red = r;
     green = g;
     blue = b;
